exercises/cloud-run/secret-manager/requirements.txt
flask
requests
gunicorn
google-cloud-secret-manager
google-cloud
